refactor: moved add_handlers functions to handlers module
